MySQL code modified for constraints

+++ src/models/MySQL/User.js
@@ -1,4 +1,6 @@
 const db = require('../../config/dbconfig/MySQLDB')
+const randomstring = require("randomstring")
+const datetime = require("node-datetime")
 
 module.exports = class User {
   getAll() {
@@ -9,17 +11,37 @@ module.exports = class User {
     return db.execute(`SELECT * FROM users WHERE id=?`, [id])
   }
 
-  save(user) {
-    return db.execute(`INSERT INTO users (username, email, password) 
-      VALUES (?, ?, ?)`, [user.username, user.email, user.password])
+  async save(user) {
+    const users = await db.execute(`INSERT INTO users (username, email, password, api_token, api_token_created_at, created_at, updated_at) 
+      VALUES (?, ?, ?, ?, ?, ?, ?)`, [user.username, user.email, user.password, randomstring.generate(), datetime.create().format('Y-m-d H:M:S'), datetime.create().format('Y-m-d H:M:S'), datetime.create().format('Y-m-d H:M:S')])
+    
+    if (users[0].affectedRows) {
+      return db.execute(`INSERT INTO profiles (fname, mname, lname, created_at, updated_at, user_id) VALUES (?, ?, ?, ?, ?, ?)`,
+        [user.fname, user.mname, user.lname, datetime.create().format('Y-m-d H:M:S'), datetime.create().format('Y-m-d H:M:S'), users[0].insertId])
+    } else {
+      return false
+    }
   }
 
-  update(user, id) {
-    return db.execute(`UPDATE users SET username=?, email=?, password=?
-      WHERE id=?`, [user.username, user.email, user.password, id])
+  async update(user, id) {
+    const users = await db.execute(`UPDATE users SET username=?, email=?, password=?, updated_at=? WHERE id=?`, 
+      [user.username, user.email, user.password, datetime.create().format('Y-m-d H:M:S'), id])
+    
+    if (users[0].affectedRows) {
+      return db.execute(`UPDATE profiles SET fname=?, mname=?, lname=?, updated_at=? WHERE user_id=?`,
+        [user.fname, user.mname, user.lname, datetime.create().format('Y-m-d H:M:S'), id])
+    } else {
+      return false
+    }
   }
 
-  delete(id) {
-    return db.execute(`DELETE FROM users WHERE id=?`, [id])
+  async delete(id) {
+    const users = await db.execute(`DELETE FROM profiles WHERE user_id=?`, [id])
+    
+    if (users[0].affectedRows) {
+      return db.execute(`DELETE FROM users WHERE id=?`, [id])
+    } else {
+      return false
+    }
   }
 }
